Check for "GET" in Behavior....IsGetOperation
In NewtonsoftJsonBehavior.IsGetOperation(OperationDescription operation)
the check to determine whether an operation was a "GET" operation was
missing [WebInvoke(Method="GET")] WCF OperationContracts.  This could be
an oversight on my part.  However, with WCF I've always tended to use
WebInvoke over WebGet.  This could be a deficiency on my part.  However,
I figured I would suggest this fix in the event that anyone else does
WCF in this potentially non-orthodox fashion.

Jeroen, Thank you for this BehaviorExtension to employ NewtonsoftJson.
Like you I prefer the date format as well. Not to mention the potential
performace benefits of Json.NET vs. the data contract serializer...